bug fixed (ish)
unfortunately, the only way to fix this that i could figure out was to
remove the react-window that virtualizes the global sequencer array map.
the window that virtualizes the nodes within each sequencer is working
as before and does provide a significant performance increase so I will
call this a win.